slack/requirements.txt
slack-bolt==1.11.0
openai>=1.3.0
python-dotenv>=1.0.0
